backend/requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.3.2
dj-database-url==2.2.0
Django==5.1.1
django-cors-headers==4.4.0
django-ratelimit==4.1.0
django-rest-framework==0.1.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
Faker==28.4.1
gunicorn==23.0.0
idna==3.10
packaging==24.1
pillow==10.4.0
psycopg2-binary==2.9.9
PyJWT==1.7.1
python-dateutil==2.9.0.post0
requests==2.32.3
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
urllib3==2.2.3
whitenoise==6.7.0
